finalProject/requirements.txt
pandas==2.2.2
numpy==2.1.1
scikit-learn==1.2.2
scikit-surprise==1.1.3
matplotlib==3.7.1
seaborn==0.12.2
tqdm==4.66.1